# app/modules/users/schemas/auth.py
from typing import Optional

from pydantic import BaseModel, EmailStr, field_validator

from app.modules.users.utils.auth import validate_password


class TokenResponse(BaseModel):
    access_token: str
    refresh_token: str
    token_type: str = "bearer"
    expires_in: int


class RefreshTokenRequest(BaseModel):
    refresh_token: str


class LoginRequest(BaseModel):
    email: EmailStr
    password: str

    @field_validator("password")
    @classmethod
    def validate_password(cls, value: str) -> str:
        if not value or len(value) < 8:
            raise ValueError("Password must be at least 8 characters")
        return value


class RegisterRequest(BaseModel):
    email: EmailStr
    password: str
    name: Optional[str] = None
    avatar_url: Optional[str] = None
    bio: Optional[str] = None
    position: Optional[str] = None

    @field_validator("password")
    @classmethod
    def validate_password(cls, value: str) -> str:
        return validate_password(value)


class AuthResponse(BaseModel):
    user: dict
    token: TokenResponse